Refactored to proper java project
So the demo code provided by my teacher + my own generated code using antlr did not seem to work. I since then refactored the codebase to a proper java project using the Create java project command in vscode.

This seemed to not work at first (kept getting the same import errors), but after doing some more research i found that adding a proper package declaration in the files generated by antlr everything worked. I wrote a quick jupyter notebook to automate the appending of the package decleration to any java file in a subdirectory of src and if there is a grammer file with the same starting name in that directory.